requirements.txt
Flask
PyJWT
Flask-SQLAlchemy